requirements.txt
django
djangorestframework
pymongo
psycopg2
